refactor(useCombatLoop): extract useCombatTick
Pulls the 50ms engaged tick, 10s HP sync ticker, usePotion callback, and
all player-vitals state (playerHp, barrier, leech, dead, swing progress,
swing index) into a dedicated hook. The state machine hands in `enemy`
(prop) plus `resolveKill` + `updateEnemy` + `pushEvent` callbacks.

Vitals reset (zone activation) and the deactivation HP-sync flush live
inside the tick hook. The orchestrator's activation effect now only
resets state-machine state (state / enemy / lastKill / bossIntroStage /
pendingIncenseRef).

Fresh-spawn reset of swing progress refs is gated on `enemy?.scaled`
identity — stable within a fight, fresh across spawns. usePotion's deps
drop `playerHp` (read via ref) so the callback identity doesn't churn
at 50ms cadence.

useCombatLoop.ts drops from 774 to 433 lines.
