+++ src/hooks/useCombatTick.ts
@@ -1,0 +1,406 @@
+// ─────────────────────────────────────────────────────────────────────────────
+//  Engaged-state combat loop: 50ms tick + 10s periodic HP sync + the potion
+//  consumable. Owns player vitals (HP, barrier, leech, dead) — the
+//  orchestrator hands in the state machine's `state` + `enemy` and the
+//  `resolveKill` callback that handles state transitions on death.
+//
+//  Damage flow per tick:
+//    leech heal → barrier recovery → player swing → enemy swing → thorns
+//
+//  Mutations fire on:
+//    - usePotion (consumePotion, optimistic decrement on the mutation hook)
+//    - HP sync ticker (syncHp every 10s when hp changed; gated on deadRef)
+//    - deactivation (syncHp flush, only when not dead)
+// ─────────────────────────────────────────────────────────────────────────────
+
+import { useMutation } from "convex/react";
+import { useCallback, useEffect, useRef, useState } from "react";
+import {
+	type BarrierState,
+	damageBarrier,
+	makeBarrierState,
+	rescaleBarrier,
+	tickBarrierRecovery,
+} from "#/game/combat/barrier";
+import { POTION_HEAL_FRACTION } from "#/game/combat/constants";
+import {
+	applyDamageToBarrierThenLife,
+	rollEnemyAttack,
+	rollPlayerSwing,
+} from "#/game/combat/damage";
+import type { LeechInstance } from "#/game/combat/leech";
+import { createLeechInstance, tickLeechInstances } from "#/game/combat/leech";
+import type { ComputedCharacterStats } from "#/game/stats/types";
+import {
+	applyCharacterDelta,
+	findCharacter,
+} from "#/lib/optimistic-character";
+import { playSfx } from "#/lib/sfx";
+import { api } from "../../convex/_generated/api";
+import type { Id } from "../../convex/_generated/dataModel";
+import type { Enemy } from "./useCombatLoop";
+import type { DamageEvent } from "./useDamageEvents";
+import { useTicker } from "./useTicker";
+
+const TICK_INTERVAL_MS = 50;
+// Periodic sync is insurance against a mid-combat refresh — the deactivation
+// effect (retreat / view change) already flushes the latest HP synchronously
+// on graceful exits. 10s of potential lost-on-refresh HP is the tradeoff for
+// the call-volume reduction. Self-skips when HP hasn't changed since the
+// last write, so idle players make zero calls regardless of this interval.
+const HP_SYNC_INTERVAL_MS = 10000;
+
+type Params = {
+	characterId: Id<"characters">;
+	active: boolean;
+	// True only when state === "engaged" — the tick gate. The state machine
+	// passes its current state so the tick can stop firing during boss intro
+	// / victory / acampamento.
+	isEngaged: boolean;
+	enemy: Enemy | null;
+	stats: ComputedCharacterStats;
+	initialHp: number;
+	potions: number;
+	onPlayerDeath: () => void;
+	// Reset to "victory" by resolveKill before this returns — used by the
+	// tick to fall through after a player-swing kill or thorns-reflect kill.
+	resolveKill: (killed: Enemy) => void;
+	pushEvent: (event: Omit<DamageEvent, "id">) => void;
+	// Mutates `enemy` when a swing connects (mid-tick HP write). The state
+	// machine's enemy ref is the source of truth.
+	updateEnemy: (next: Enemy) => void;
+};
+
+export function useCombatTick({
+	characterId,
+	active,
+	isEngaged,
+	enemy,
+	stats,
+	initialHp,
+	potions,
+	onPlayerDeath,
+	resolveKill,
+	pushEvent,
+	updateEnemy,
+}: Params) {
+	const maxHp = stats.maxLife;
+
+	const [playerHp, setPlayerHp] = useState(initialHp);
+	const [barrier, setBarrier] = useState(() =>
+		makeBarrierState(stats.maxBarrier),
+	);
+
+	const playerHpRef = useRef(playerHp);
+	playerHpRef.current = playerHp;
+	const initialHpRef = useRef(initialHp);
+	initialHpRef.current = initialHp;
+	const lastSyncedHpRef = useRef(initialHp);
+	const barrierRef = useRef(barrier);
+	barrierRef.current = barrier;
+	const leechRef = useRef<LeechInstance[]>([]);
+	const deadRef = useRef(false);
+
+	// Combat-progress refs: how far each side is into its next swing (in
+	// "swings", incremented by dt*rate per tick; ≥1 fires the swing).
+	const playerProgressRef = useRef(0);
+	const enemyProgressRef = useRef(0);
+	// Index into stats.swings — advances modulo swings.length so dual-wield
+	// alternates main / off hand.
+	const nextSwingIndexRef = useRef(0);
+
+	const syncHp = useMutation(api.combat.syncHp);
+	const consumePotion = useMutation(api.combat.usePotion).withOptimisticUpdate(
+		(localStore, args) => {
+			const char = findCharacter(localStore, args.characterId);
+			if (!char) return;
+			applyCharacterDelta(localStore, args.characterId, {
+				potions: Math.max(0, (char.potions ?? 0) - 1),
+			});
+		},
+	);
+
+	// Keep barrier max in sync with the stat engine. Gear swaps mid-combat
+	// rescale rather than reset to full.
+	useEffect(() => {
+		setBarrier((prev) => rescaleBarrier(prev, stats.maxBarrier));
+	}, [stats.maxBarrier]);
+
+	// Fresh spawn → reset progress refs so the first swing fires at the same
+	// cadence as a from-zero fight. `enemy?.scaled` identity is stable through
+	// a single fight (damage swaps `currentHp` but keeps `scaled` by reference)
+	// and changes only when a new monster is rolled. INVARIANT: every code
+	// path that creates a new `Enemy` MUST allocate a fresh `scaled` object —
+	// reusing or mutating in place would silently stop resetting swing
+	// alternation across fights.
+	useEffect(() => {
+		if (enemy) {
+			playerProgressRef.current = 0;
+			enemyProgressRef.current = 0;
+			nextSwingIndexRef.current = 0;
+		}
+	}, [enemy?.scaled]);
+
+	// Activation: reset vitals on zone entry. Deactivation: flush HP sync if
+	// the player is alive (graceful retreat / view change).
+	const activeRef = useRef(active);
+	useEffect(() => {
+		const wasActive = activeRef.current;
+		if (active && !wasActive) {
+			playerHpRef.current = initialHpRef.current;
+			setPlayerHp(initialHpRef.current);
+			lastSyncedHpRef.current = initialHpRef.current;
+			deadRef.current = false;
+			leechRef.current = [];
+			nextSwingIndexRef.current = 0;
+		} else if (!active && wasActive && !deadRef.current) {
+			syncHp({ characterId, hpCurrent: playerHpRef.current }).catch(() => {});
+			lastSyncedHpRef.current = playerHpRef.current;
+		}
+		activeRef.current = active;
+	}, [active, characterId, syncHp]);
+
+	const enemyAttackSpeed = enemy?.scaled.attackSpeed ?? 1;
+	const tickRate = stats.tickRate || 1;
+	const hasSwings = stats.swings.length > 0;
+
+	useTicker(active && isEngaged && enemy !== null, TICK_INTERVAL_MS, () => {
+		if (deadRef.current) return;
+		const currentEnemy = enemy;
+		if (!currentEnemy || currentEnemy.currentHp <= 0) return;
+
+		const dt = TICK_INTERVAL_MS / 1000;
+
+		// Leech ticks every frame regardless of swing timing.
+		if (leechRef.current.length > 0) {
+			const { healed, instances } = tickLeechInstances(
+				leechRef.current,
+				dt,
+				maxHp,
+			);
+			leechRef.current = instances;
+			if (healed > 0 && !deadRef.current) {
+				const next = Math.min(maxHp, playerHpRef.current + healed);
+				if (next !== playerHpRef.current) {
+					playerHpRef.current = next;
+					setPlayerHp(next);
+				}
+			}
+		}
+
+		// Barrier recovery timer ticks too.
+		if (
+			barrierRef.current.recoveryRemaining !== null &&
+			barrierRef.current.recoveryRemaining > 0
+		) {
+			const next = tickBarrierRecovery(barrierRef.current, dt);
+			if (next !== barrierRef.current) {
+				barrierRef.current = next;
+				setBarrier(next);
+			}
+		}
+
+		playerProgressRef.current += dt * tickRate;
+		enemyProgressRef.current += dt * enemyAttackSpeed;
+
+		const playerSwing = playerProgressRef.current >= 1 && hasSwings;
+		if (playerSwing) playerProgressRef.current -= 1;
+		const enemySwing = enemyProgressRef.current >= 1;
+		if (enemySwing) enemyProgressRef.current -= 1;
+
+		if (playerSwing) {
+			const swingIndex = nextSwingIndexRef.current % stats.swings.length;
+			const swing = stats.swings[swingIndex];
+			nextSwingIndexRef.current =
+				(nextSwingIndexRef.current + 1) % stats.swings.length;
+
+			const result = rollPlayerSwing({
+				swing,
+				stats,
+				defender: {
+					armor: currentEnemy.scaled.armor,
+					evasion: currentEnemy.scaled.evasion,
+					accuracy: currentEnemy.scaled.accuracy,
+					level: currentEnemy.level,
+					resistances: currentEnemy.scaled.resistances,
+				},
+			});
+
+			if (!result.isMiss && result.amount > 0) {
+				const newEnemyHp = Math.max(0, currentEnemy.currentHp - result.amount);
+				const updated: Enemy = { ...currentEnemy, currentHp: newEnemyHp };
+				updateEnemy(updated);
+				pushEvent({
+					amount: result.amount,
+					target: "enemy",
+					isCrit: result.isCrit,
+					weaponType: swing.weaponType,
+				});
+				playSfx("hit.wav", {
+					volume: 0.3,
+					pitchVariance: 0.1,
+					exclusive: true,
+				});
+
+				// MVP: leech on physical only; elemental leech is a future mod.
+				// Apply globally regardless of which weapon swung.
+				const leechSrc = result.breakdown.physical;
+				if (stats.lifeLeechPercent > 0 && leechSrc > 0) {
+					const inst = createLeechInstance(leechSrc, stats.lifeLeechPercent);
+					if (inst) leechRef.current.push(inst);
+				}
+
+				if (stats.lifeGainOnHit > 0 && !deadRef.current) {
+					const next = Math.min(
+						maxHp,
+						playerHpRef.current + stats.lifeGainOnHit,
+					);
+					if (next !== playerHpRef.current) {
+						playerHpRef.current = next;
+						setPlayerHp(next);
+					}
+				}
+
+				if (newEnemyHp <= 0) {
+					resolveKill(updated);
+					return;
+				}
+			} else {
+				pushEvent({
+					amount: 0,
+					target: "enemy",
+					isCrit: false,
+					isMiss: true,
+				});
+				playSfx("errarHit.wav", {
+					volume: 0.25,
+					pitchVariance: 0.1,
+					exclusive: true,
+				});
+			}
+		}
+
+		if (enemySwing) {
+			const attack = rollEnemyAttack({
+				enemyAccuracy: currentEnemy.scaled.accuracy,
+				physicalDamage: currentEnemy.scaled.physicalDamage,
+				elementalDamage: currentEnemy.scaled.elementalDamage,
+				defender: {
+					armor: stats.armor,
+					evasion: stats.evasion,
+					accuracy: stats.accuracy,
+					level: currentEnemy.level,
+					resistances: stats.resistances,
+					blockChance: stats.blockChance,
+				},
+			});
+			if (attack.isMiss) {
+				pushEvent({ amount: 0, target: "player", isMiss: true });
+				playSfx("esquiva.wav", {
+					volume: 0.5,
+					pitchVariance: 0.1,
+					exclusive: true,
+				});
+			} else if (attack.isBlocked) {
+				// No damage to barrier/life, but the hit still "lands" for thorns.
+				pushEvent({ amount: 0, target: "player", isBlocked: true });
+				playSfx("block.wav", {
+					volume: 0.5,
+					pitchVariance: 0.1,
+					exclusive: true,
+				});
+			} else if (attack.amount > 0) {
+				// Apply to barrier first, then life.
+				const { state: nextBarrier, lifeOverflow } = damageBarrier(
+					barrierRef.current,
+					attack.amount,
+				);
+				barrierRef.current = nextBarrier;
+				setBarrier(nextBarrier);
+
+				const result = applyDamageToBarrierThenLife(
+					lifeOverflow,
+					0,
+					playerHpRef.current,
+				);
+				playerHpRef.current = result.newLife;
+				setPlayerHp(result.newLife);
+				pushEvent({ amount: attack.amount, target: "player" });
+				playSfx("tomandoHit.wav", {
+					volume: 0.3,
+					pitchVariance: 0.1,
+					exclusive: true,
+				});
+
+				if (result.newLife <= 0 && !deadRef.current) {
+					deadRef.current = true;
+					playSfx("morte.wav");
+					queueMicrotask(() => onPlayerDeath());
+				}
+			}
+
+			// Per CONTEXT.md → Defenses → Block: thorns still reflect on block.
+			// Not on miss. If reflection kills, fall through to the same victory
+			// branch the player-swing uses.
+			if (!attack.isMiss && stats.thorns > 0 && !deadRef.current) {
+				const reflected = Math.max(1, Math.floor(stats.thorns));
+				const enemyAfter = Math.max(0, currentEnemy.currentHp - reflected);
+				const updated: Enemy = { ...currentEnemy, currentHp: enemyAfter };
+				updateEnemy(updated);
+				pushEvent({
+					amount: reflected,
+					target: "enemy",
+					isThorns: true,
+				});
+				if (enemyAfter <= 0) {
+					resolveKill(updated);
+				}
+			}
+		}
+	});
+
+	// Periodic HP sync — skips when hp hasn't changed since last write.
+	useTicker(active, HP_SYNC_INTERVAL_MS, () => {
+		if (deadRef.current) return;
+		const hp = playerHpRef.current;
+		if (hp === lastSyncedHpRef.current) return;
+		lastSyncedHpRef.current = hp;
+		syncHp({ characterId, hpCurrent: hp }).catch(() => {
+			lastSyncedHpRef.current = -1;
+		});
+	});
+
+	// Guards read the ref so the callback identity is stable across combat
+	// ticks (playerHp state changes every damage frame; including it in deps
+	// would recreate the callback at 50ms cadence and invalidate any
+	// downstream memoization).
+	const usePotion = useCallback(async () => {
+		if (potions <= 0 || playerHpRef.current >= maxHp) return;
+		const prevHp = playerHpRef.current;
+		const heal = Math.floor(maxHp * POTION_HEAL_FRACTION);
+		const optimisticHp = Math.min(maxHp, prevHp + heal);
+		const appliedHeal = optimisticHp - prevHp;
+		playerHpRef.current = optimisticHp;
+		setPlayerHp(optimisticHp);
+		lastSyncedHpRef.current = optimisticHp;
+		// HP is local-only (hook state). Potion count comes from the Convex
+		// character query; `consumePotion` wraps the localStore decrement so
+		// no race against concurrent `recordKill` drops.
+		try {
+			await consumePotion({ characterId });
+		} catch {
+			const reverted = Math.max(0, playerHpRef.current - appliedHeal);
+			playerHpRef.current = reverted;
+			setPlayerHp(reverted);
+			lastSyncedHpRef.current = -1;
+		}
+	}, [potions, maxHp, characterId, consumePotion]);
+
+	return {
+		playerHp,
+		barrier,
+		usePotion,
+	};
+}
+
+export type { BarrierState };
